test(inspections): add extensive request specs for inspections controller
- Add tests for assessments disabled scenario returning 404
- Add JSON format response test for inspection data
- Add image processing error handling test
- Add event logging verification on unit change
- Add HEAD request test for inspections
- Add inspection event log display test
- Add prefill functionality tests for inspection edit
- Add validation for tab parameter in edit action
- Add error handling tests for inspection event logging
- Add tests for tracking changes in inspection updates
- Add tests for validating inspection completability with error handling

This commit significantly improves test coverage and robustness of the inspections controller.
